backend/requirements.txt
annotated-doc
annotated-types
anyio
bce-python-sdk
certifi
chardet
charset-normalizer
click
colorlog
fastapi
filelock
fsspec
future
h11
hf-xet
httpcore
httpx
huggingface_hub
idna
imagesize
lxml
modelscope
numpy
#opencv-contrib-python
packaging
pandas
pdf2image
pillow
prettytable
psutil
py-cpuinfo
pyclipper
pycryptodome
pydantic
pydantic_core
pypdfium2
pytesseract
python-bidi
python-dateutil
python-docx
python-multipart
PyYAML
requests
ruamel.yaml
setuptools
shapely
shellingham
six
starlette
tqdm
typer-slim
typing-inspection
typing_extensions
ujson
urllib3
uvicorn
wcwidth
pydantic-settings
motor
minio
sentence-transformers
chromadb
pypdf2
python-jose[cryptography]
passlib[bcrypt]
tenacity
PyMuPDF
#fitz
#bson